+++ src/utils/colors.js
@@ -1,0 +1,16 @@
+const colors = {
+  primaryDark: '#00b197',
+  primaryLight: '#53cab6',
+  primary: '#51cbb4',
+
+  backgroundLoading: '#f3f3f3',
+  backgroundLoadingDark: '#ecebeb',
+  greyIcon: '#b7c1ca',
+  backgroundSearch: '#f7f7f7',
+  titleColorText: '#2f3a4e',
+  colorText: '#737f97',
+
+  white: '#FFFFFF',
+  black: '#000000',
+};
+export default colors;