+++ CHANGELOG.md
@@ -17,3 +17,12 @@
 ## [ADD] 2019-11-27 02:24
 
 - Add text changes in Changelog
+
+## [ADD, CHANGE] 2019-11-28 20:09
+
+- Add all genres of anime/manga
+- Add genre component
+- Add data card(anime/manga)
+- Add default image to data card
+- Add pagination to genre and data card screen
+- Change UI of Home screen